requirements.txt
# GPF Reboot Requirements - Modern Environment
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter>=2.1.0
torch-sparse>=0.6.17
torch-cluster>=1.6.1
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.62.0
pyyaml>=6.0
matplotlib>=3.5.0
